requirements.txt
torch>=1.10.0
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
mlflow>=1.20.0
loguru>=0.5.3
pyyaml>=6.0
jinja2>=3.0.0
tqdm>=4.60.0
notebook>=6.4.0